% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_usd_to_eur.R
\name{convert_usd_to_eur}
\alias{convert_usd_to_eur}
\title{Takes the provided number as USD and converts it to EUR on the current exchange rate}
\usage{
convert_usd_to_eur(usd, decimals = 5)
}
\arguments{
\item{decimals}{number of decimals for the exchange rate}

\item{x}{number of USD}
}
\value{
integer
}
\description{
Takes the provided number as USD and converts it to EUR on the current exchange rate
}
\examples{
convert_usd_to_eur(5)
convert_usd_to_eur(42)
}
