% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eur.R
\name{eur}
\alias{eur}
\title{Provided value in € format, separated by comma every 3 digits, rounded to 2 digits}
\usage{
eur(value)
}
\arguments{
\item{amount}{amount to transform to €}
}
\value{
string
}
\description{
Provided value in € format, separated by comma every 3 digits, rounded to 2 digits
}
\examples{
eur(87654321)
eur(-632678)
}
\seealso{
uneur()
}
