% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusd.R
\name{eurusd}
\alias{eurusd}
\title{Most recent USD/EUR exchange rate as number}
\usage{
eurusd(decimals = 5)
}
\arguments{
\item{decimals}{number of decimals to show}
}
\value{
integer
}
\description{
Most recent USD/EUR exchange rate as number
}
\examples{
eurusd()
eurusd(1000)
}
