% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusds.R
\name{eurusds}
\alias{eurusds}
\title{Daily exchange rates between the provide dates}
\usage{
eurusds(date_from, date_to, decimals = 5)
}
\arguments{
\item{date_from}{starting date of the period, back-filled for missing dates}

\item{date_to}{ending date of the period, back-filled for missing dates}

\item{decimals}{number of decimals to show}
}
\value{
integer
}
\description{
Daily exchange rates between the provide dates
}
\examples{
eurusds('2019-06-01', '2019-06-08')
}
