% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneur.R
\name{uneur}
\alias{uneur}
\title{Converts currency string into sole number}
\usage{
uneur(eur)
}
\arguments{
\item{eur}{€ value}
}
\value{
integer
}
\description{
Converts currency string into sole number
}
\examples{
uneur('€99,999.338309475')
uneur('-€987,765,333.4657')
}
\seealso{
eur()
}
